JANITORIAL: Fix SDL_revision.h generation
Remove the committed SDL_revision.h to prevent it always showing up as
modified.  Instead, generate the default contents in updaterev.sh as
needed.  Modify showrev.sh to generate a better commit id, which includes
a commit sequence number.  Modify updaterev.sh to be able to extract this
number for the SDL_REVISION_NUMBER define.